refactor: 추천 도서 페이지 UI 수정
